Fixed: IPA entries not appearing in pronunciation sections
When a pronunciation section had both a template, like hyphenation, that
generates data that goes into the `prefix` variable, that prefix was never
reset within the loop where we go through the data in the pronunciation
section and so subsequent loopings would get garbage data. Apparently if
`sounds` > {`ipa`: something} also has a `form` field in the object it
gets rejected at some point.

This was fixed by moving `prefix` from outside the scope of the for loop
to the inside, resetting it each loop. "current_pos" seems like it should
be outside the scope, as is the (currently unused) have_pronunciation
debug print variable.