Use parens () as delimiter in zh-pron pronunciation values
Sometimes you have a line like:
>    (Pinyin): lǜ, lù (lü4, lu4)

where the contents of the parentheses are just alternatives to the others.
These are broken up as the commas (or " / " if there was any), so the
values are "lü", "lù (lü4" and "lu4)".

Using the parens as delimiters should not break things more than they are
already broken: values separated by the () seem like they should already
be separated, or if they contained non-pronunciation values they don't
belong anyhow, so it's just broken in a different way.