zh-pron "toneless final syllable variant" handled
Made a portion of parse_expanded_zh_pron into a subfunction,
and handled pron-values with an arrow separately as a special
case.

generate_pron() returns a `pron` value, which can now be None,
so you need to check if it exists. It will return "None" if
the "ipa" and "zh-pron" fields are both empty.