Fixes #221, broken Portuguese pronunciation section
Issue was in flattened_tree(), which was a kludgey way
to ignore irrelevant pronunciation hierarchies.

It presupposed to everything should be inside a LIST or
LIST_ITEM node (where it would then yield the contents),
but if the pronunciation section doesn't actually start
with a * then it would skip contents, like a template
actually containing the pronunciation tables.

Fixed by yielding the node as a last resort.