Create new template override functions
There are certain templates, like egy-glyph, which are difficult
to handle after they have been expanded for one reason or another.
In the case of egy-glyph, it creates a whole table that is "inline"
on the same line as other items, and uses images which return no
text by default.

The simplest way to handle these is to to always intercept them
and transform them into formatted data/text.

In the case of egy-glyph and egy-glyph-img and return their arguments
as a formatted string. egy-glyph-img is just turned into its
hieroglyphic number code inside «» style quotation marks, and
the same, in general, done with egy-glyph.

For more about how this is actually done, see the commit in
wikitextprocessor where we introduce the changes into core.py
to enable this template-interception.