ctx -> wxr.wtp, config -> wxr.config
Creates a Wiktextract-specific context for each process.

Instead of misusing Wikitextprocessor's Wtp context (`ctx`),
these commits consolidate ctx and config under wxr, which
is also meant to be used as the general wiktextract context.

wxr.wtp replaces what was previously ctx and wxr.config replaces
the config object.

This was done in this way make the transferring easier, so that
there wouldn't be need to actually touch Wtp and WiktionaryConfig yet.
Now with the specific wxr-dot syntax, it's 500% easier to make big
search and replace changes for different fields in config, for example,
so consolidating everything under wxr should be easy.

The name 'wxr' was chosen just so that it's easy to search and
replace later, if needed. It's also short.

wxr will also, in the future, contain the general parameters
`word`, `lang` and `pos`, which are repeated extremely commonly
in different function signatures and which are suitable to be
part of the context object because they *are* the context of
parsing a page.

Other future plans include creating task-specific context for
files like inflection.py etc, where the number of parameters
functions have tend to grow too big.