Fixed issue with pronunciation templates
Some templates, like la-IPA (praemitto) generate several lines
of pronunciation data; for some reason, this was not caught by
other kludges related to other pron-templates like this
(probably because the template was part of a LIST_ITEM), so the
parsing of the multi-line data was a bit broken in regards to
tags. nihil/Latin is fine because the template as used there
only generates single lines by coincidence.

This is fixed by interjecting a generator function in-between
flattened_tree and the parsing loop that expands the node with
clean_node outside of the loop (instead of inside it) and then
splits the text by lines.