--override now accepts directories too
Override is more of a debugging tool, but needs must (that's an actual phrase,
look it up).

In case a persistent (but hopefully temporary) override of a module is
required, --override now accepts a directory as a path, and will use all files
(not subdirectories) of that directory.

This way, you can add a module to this repo inside a override directory and
give an "agnostic" --override parameter without having to fiddle with the
parameter every time you remove or add new modules to be overwritten.

This patch was created because we have yet another instance of kaikki.org
getting inundated with Lua errors (halting the regen because of the sheer
number of errors crossing the allowable threshold). Because this is a universal
problem with en.wiktionary, it might as well be fixed in wiktextract:

- copy an older functioning version of the module that breaks

- put it in a override folder

- call wiktwords in kaikki's regenerated script with --override

- remove the module from the override folder when the edit water isover