post_template_fn for pre-expanded multitrans
Issue: when we have translation sections with:

```
{{trans-top}}{{multitrans|data=
* Foobarian translation
* Barfoorian translation
}}
{{trans-bottom}}
```

the multitrans gets pre-expanded and you end up with:

```
{{trans-top}}* Foobarian translation
* Barfoorian translation
```

when parsing, so the first item of the list is misparsed as some
random text (= the senses of the translation) and breaks the
list.

The post_template_fn functions added here just add a newline
to each multitrans template expansion that doesn't have one,
so that the output is:

```
{{trans-top}}
* Foobarian translation
```

which parses correctly. This isssue is caused by the way we
pre-expand some templates and not all. Pre-expanding trans-top
doesn't seem to work either, and I can't understand why.