Pre-expand translation section templates
There are too many issues with how multitrans interacts and breaks with
trans-top, trans-bottom, checktrans-top, and checktrans-bottom because of how
multitrans is being pre-expanded before these others, creating broken tables
and divs when the order of parsing gets wonky.

I've recently added some code that pulls 'sense' data from trans-top, but
otherwise this kind of attribute-use is not used here too much, and even if it
is present in the code we still have text-parsing versions of things that
should do the job.

This mainly affects multitrans when it encompasses and intertwines
with several sense sections. It just breaks down if we don't
pre-expand everything together with multitrans.