"Contraction" is now its own part-of-speech
Sometimes, an entry fits best under "contraction" like 'I've'
(which doesn't fit as a Noun or a Verb, maybe Phrase),
and these are reasonably common, so I've taken them away
from `abbrev` (which is an umbrella pos-tag for a number
of things) and added `contraction`.

Noticed that when copying data from parts_of_speech to
pos_subtitles.json in the data folders the migration was possibly not
completed. Will take a look after consulting next week.