Testing overriding Module:xnt-decl
This module might be triggering an annoying bug.

In it this module, `table` is redefined as a global
string (obviously by accident), which is corrected
here by making it local and a different name.

The problem is that the global `table` seems to be
polluting the globals, even when those globals
should be reset by our Lua compatibility stuff.

In the files in pages/Module/* only xnt-decl
redeclares `table` as a global variable with the pattern
`table =` in the code, and the first instance of
the languge Narragansett appearing in extract.log
for the kaikki regen (which goes through all languages)
is at 22% of the file, which is the point where
the bug starts to appear in other cases. This is
pretty conclusive evidence that `table` is being
overwritten.