Fix(?) weird EXCEPTION
When running the kaikki regen, there's a weird EXCEPTION in extract.log
that basically ends as a regex pattern error inside quote_kept_ruby(),
but I can't recreate this error with the individual page in any
language combination (English, Japanese, Polish or all of them).

The regex error seems to be a "can't have multiples of an empty string"
error caused by how the formatted string gets values from "|".join(),
which *can* be empty if the values fed to it are empty.

This seems to indicate a deeper bug, and like with all deeper bugs that
are only present when doing batch parsing, I suspect the culprit is
something breaking in global namespace stuff.