Create new Wiktextract-specific context object
WiktextractContext is a new wrapper object for
Wtp ("ctx") and WiktionaryConfig (config) that
has those two as its fields.

Wtp should not be used a context object for things
inside Wiktextract, and there are too many functions
with really long parameter lists that need to be
trimmed down, at least a little bit, using
a context object holding them; things like what
language, word and part of speech is being currently
parsed.

Currently, the only changes are the creation of
WiktextractContext and switching out all calls and
references to `ctx` and `config` so that they are
through `wxr.wtp` and `wxr.config` instead.

WiktextractConfig mainly slots in at the same place
WiktionaryConfig does, so it would have been possible
to just change WiktionaryConfig into a context object,
but this is more straightforward and making changes
to `wxr.wtp` and `wxr.config` is easier to automate.