Testing Lua global reset issue with old xnt-decl
This is just to see whether xnt-decl breaks `table`;
hopefully it doesn't, which means a major issue with
the global namespace not resetting properly has been
fixed.

Annoyingly, this is best tested by just letting
a whole new batch run, because the issue is hard to replicate
with sufficient fidelity otherwise.