Add --search-pattern command line parameter
--search-pattern allows the user to add a "body LIKE ?"
pattern when doing big batches.

Functions currently only when used with wikt-db.

By supplying a SQL text pattern, where % is zero, one
or more wildcard characters and _ is exactly one,
you can process only pages that contain that pattern;
for example "%==Old_English==%" would find pages with
a header "Old English" or "Old-English". However,
as you can see from the example, this cannot be
100% relied on to deliver pages with specific languages,
unless the data can be trusted to be consistent.
So searching "English" should be reliable, but a
language where the name has variants is not. Some
editions of Wiktionary also use different methods,
like templates, to generate language headers.

It is better to have a ton of false positives, because
processing them does not cause problems, only
slows down the process because those pages get parsed
and an attempt is made to find the correct languages
within those pages. The SQLite "LIKE" process is also
ridiculous fast. Genuinely super-fast.