Fixed issue with Module:columns
Removed columns.txt from override/

The issue was a change from using "* list item"
syntax in the Module:columns output to using straight-up
HTML elements (ul and li).

Seeing as how <li> elements are equivalent to `* item`,
these should be handled in a similar way, and happily
it seems that I could just copy paste the code from
one place and add it in the HTML portion of the
recursive content crawler in parse_linkage_recurse.