Fixed issue with a "demoted" defaultdict (->dict)
In thesaurus, I'd changed a dict from defaultdict to
a normal dict because of issues with None that were
being expressed as json `null` (which broke things)
and missed one line where we appended to a list
that was not necessarily there.

Fixed by checking for existence of list and creating
an empty one if needed.