Changed data outputted as list into string
The json data we output should really not have fields that alternate between
different types: if a field in a specific location can be either a string or a
list, that causes problems downstream, like it has for me these past few
days...

Kaikki.org, our own data presentation platform, has failed due to two different
different bugs becaus of this; the first was that it assumed the "roman" field
in a certain place would always be a string. The second was that our mapping
script (jsonmapping.py in the kaikki repo) doesn't have the capability to
create different type versions of nodes it sees (which could be viewed as an
oversight, but in our case we could also see it as a way to enforce type
conformity).

As a rule of thumb, a field that is in the singular, like "roman", should be a
string or dict. Fields in plural, like "tags" or "topics" or "senses" are
lists.

The data should not contain nil/None!