Capture URLs in linkages (synonyms, derived, etc.) instead of outputing text
Naked urls (that are changed into `[1]` form by wikitext but not us
were being captured as word text in synonyms and then splattered
all over the `synonyms` field when they were later split on
slash (I assume).

This introduces a new field `urls` in linkage data just so that
we have somewhere to put the stuff. Naked urls are ignored in the
text, and only saved in `urls`, urls with text have their text
portion stay also in the text while the url is extracted for
consistency.