Add `overrides/zh.json` file
This new JSON file contains title and body of the overwitten
pages. And all overwite data will be contained in a single file so
they won't be used in other Wiktionary dump files.

The template `語言標題` template is used in some old templates
that expand to language subtitles, but it creates `<h2>` tag not
wikitext title.