Only call `process_categories()` for the English Wiktionary dump file
This function adds categories from the last sense data to all the
previous data then removes them if the category text starts with a
different language name. But this would only kind of works for English
Wiktionary. In French Wiktionary, language name could be at the end
and usually in lower case. For example: "Interjections en français"

Most category links are created by tempaltes and are not added at
the end of page, therefore `process_categories()` would add category
data that only belong to the last word but not the entire page. And it
alo builds a huge language regex pattern. I think it's better to only
run this function for English Wiktionary for now.