Changes to how glosses and lists are handled
Should fix #314

These changes mainly clean up the two parallel lists of
heads and lists that has been collected, removing empty
pairs and then combining separated pairs.

Separated pairs can be created when there are too many
newlines in the text, and the code thinks there's a new
paragraph; this can be caused by something like people
using <br /> tags for "tighter" formatting, like in the
above issue, but this fix is more general.

The logic is: if we see a list, and there was a previous
head node that *didn't* have a parallel list, then
just replace the empty pair with this list and then continue
processing the pairs.

This also fixes an issue with heads without glosses: if
there's a head without a gloss, previously we just didn't
handle that head and skipped it. However, it can be
important to keep those, so we do now, and hopefully we
can figure out how to filter out garbage later.

This led to a test breaking that was testing how well
we handle it when the glosses are not in LISTs, but just
plain text. This was eventually fixed by adding an if-else
block (otherwise the heads are handled twice, and we generate
double entries, which was what broke the test).