Preliminary optimization to decode_tag
Some of the algorithms and code in decode_tag has exponential
in nature, which could cause huge spikes in processing time
when the function was given malformed input (like whole translation
tables or other stuff).

Also disabled a piece of code that broke the parsing of tags
containing both spaces and "/", meant to make "something/other"
parsing simpler without having to make pairings for each thing
that is encountered, but alas.

The first optimization is to cull a bunch of path candidates
when iterating over possible tags, the second is to just recognise
a pathological case and jump out of it when there are just too many
different tagsets (see "Leyte")