Handle taggable text with slashes ("/")
Previously I had in place a kludge in decode_tags that
turned some slashes into spaces so that stuff like
"nominative/plural" (which isn't a key we're familiar with)
turns into "nominative plural" and doesn't result in an
error tag.

However, this would break keys with slashes, like the
existing "masculine/feminine", which would get the space
because "masculine" and "feminine" were both valid tags.
(In this case it didn't matter, but in others it does).

This fix basically creates a wrapper function around
decode_tags (which is now `decode_tags1`) which
retries handling the source text given to it if there
are any error-tags in the return value from decode_tags1.

We use a regex of valid keys with slashes in them to
split the source text; every even string in the list
is a "candidate" for slash-replacement, while every
other string is a key that is left alone.

Retry decode_tags1 with a new source, and if we have
as many or less errors, use that return instead of
the original.