Test for </tag> not closing list
Due to the structure of our parser, HTML tags are
kind of fragmented and can't be trusted (because they
can appear from templates).

One minor issue is that in Wikitext, an HTML element
should not break a list:
========
* This is <div>

</div>
** still the same list!

* While this one

* is two different lists
========

We don't have the capability (yet?) to allow multiline
elements like the example above, *but* as a special
case we can slip through this one:

======
* This is</div>
</div>
** still the same list!
======

This is because our code breaks up lists when encountering
stuff on the very beginning of a line, and we can have
ending </tags> be special in that they don't close lists down.

However, anything else, like text, newlines and other stuff
that closes a list still closes a list, even if it is inside
an HTML element. This is because we cannot parse the HTML
prematurely, because the HTML *might not be there yet*, because
it could be generated, partially, in different templates,
for example! Why does it work in Wikimedia stuff? Because
Wikimedia doesn't care about the semantic structure of
the parse tree, it doesn't need to know that "this bit of
text is generated by this template" etc., it just expands
the templates, generating the final HTML and *then* parses it.

We're trying to parse the text while we're still assembling it;
Wikimedia assembles the text and then parses it.