Process French Wiktionary "ja-exemple" and "zh-example" templates
Because these two templates are expand to a list, they are not child
of a gloss list item, thus need to find the template node in
`process_pos_block()` after the gloss node is processed by
`extract_gloss()`.

Also limit the inflection templates to `fr-accord-*` and `fr-rég*`,
because there are many template names start with language code but
don't expand to inflection table, like "ja-exemple" is for example,
and "zh-mot" is a form line.