Check for "or" in title for special handling
Articles like "true or false" had `canonical` entries
due to the 'or' making the parser thing "true" and "false"
are alternative forms of 'true of false'.

The fix was to just add to the condition that already did
what we wanted: process text in such a way to keep the
title intact. This was previously put in place to handle
titles with commas, maybe by me? so leaving the 'or'
out was just oversight.

This fix doesn't actually change what the content of the
`canonical` fields are because they are never created
in the first place; `canonical` is for when something
we encounter in the head looks like it's the 'real' or
an alternative form of what is found in the page title.
Because the fix now results in "true or false" being
"true or false", the same as the title, it doesn't get
entered as a canonical entry.