Fixed bug in previous commit ("or" -> " or ")
Fixed a bug last week, but messed up by looking for
"or" strings instead of " or ", which would break
pages with "or" like "intueor", by causing an old
bug to manifest that hadn't appeared before (string
concatenation with + and None, apparently `v` and
`ending` in `(v + ending)` were never None before).

Added a couple of tests this time.