Add languages.json and its generator script for the German Wiktionary.
There is no authorative list of languages in the German Wiktionary
within the dump file itself. The list of languages is generated from
the page "Hilfe:Sprachkürzel" and is probably incomplete. Until another
source is found, this is the best we can do. The list will then need to
be manually completed.