Add page and gloss module for German extractor
The page module parses a page from the German Wiktionary and extracts
the language of each language entry, the pos of each pos section and
fixes the subsection hierarchy therein.

The gloss module extracts the glosses from the relevant section. It's
still quite crude and doesn't capture all subtleties of gloss entries.