Remove "multitrans" kludge with newlines
Commit f0d0195e5833e828530fcbb1854804803c6701e7 in wikitextprocessor
should fix this issue with newlines 'missing' at the start
of {{multitrans}} templates.

What was the issue, it turns out, that those phantom newlines
that I thought was something more nebulous was simply...
Missing newlines! In wikitext, when expanding a template (or
parser function), if the result text has a special character
(*;:# or "{|") at the start of the string, it inserts a
newline before the string.

This wasn't very well documented and Tatu missed it.