parse_alt_or_inflection_of: "Western-Armenian form of"
խօսիլ/Armenian

If you have a gloss that contains a valid tag with the
"dialect" value, like `("Western-Armenian", "form-of")`,
this "form of" should be interpreted as producing
"alt-of" instead.

It is unlikely, but possible, that there are glosses like
"English accusative form of", which should take form-of,
but unlikely; other variations are possible, too. But
in general, having a dialect tag indicates an alt-of.