Move code out of for loop that didn't belong there
Fixes an issue where the index of `audios` was out of range
because we emptied it in an earlier iteration of the loop.
(Line 566 with the IPA audio file stuff, `audios` is emptied
on line 639).

This was my fault from last year, for some reason this
bit of the code that should be handled just once was put
inside a large for loop, and some kind of brainfart
must have made me think that "this should run only once,
let's prevent the if-block from happening by emptying
`audios` after we've done the processing"... And I
didn't think that `audios` was accessed elsewhere in the
loop, which triggered this pretty rare bug.