Extract glosses from Russian Wiktionary
This extracts raw_glosses and attempts to extract clean glosses (without notes and tags). Currently only some tags are recognized and moved to the tags field. The rest of the tags are left in the gloss.

This work is a contribution to the EWOK project, which receives funding from LABEX ASLAN (ANR–10–LABX–0081) at the Université de Lyon, as part of the "Investissements d'Avenir" program initiated and overseen by the Agence Nationale de la Recherche (ANR) in France.