Add two new edition configuration options
French Wiktionary has a unique "Conjugaison" namespace that contains
conjugation tables. The new options allow each edition to save and
process pages according to their needs.

I don't know whether we need the "Category" namespace. There are
845761 category pages in the English dump file. The first stage will
be faster and the db file size will be smaller if we don't save this
namespace.