Remove unused global initialization function
For some reason HEAD_TAG_RE was being initialized in a specific
function for it, without any apparent reason; 1 interaction
with HEAD_TAG_RE in the codebase, no branching, no changes
to HEAD_TAG_RE. Most probably remnants of some removed logic
or an to-do that was forgotten or became unnecessary.

This was previously touched by xxyzz to fix a but in the
regex, but the weird initialization function is from earlier.

A test was modified.