Fixed minor bug with Thesaurus aliases
This code used ns_title_prefix_tuple to get a list of
aliases for 'Thesaurus', but then use a constant w[10:]
to remove "Thesaurus" only from the start of `w`.

This might never have actually triggered, if there
are now aliases for "Thesaurus", though!

The more correct way to fix this is to iterate over
the return from ns_title_prefix_tuple, but a less
correct (but probably also correct) would be to just
(revert back to?) just checking w.startswith("Thesaurus")...