Fix a stupid bug regarding "sounds" and pos
Is there anything wrong than seeing a piece of code, thinking to yourself
"that's not right, that can't work?", searching for a keyword in there to
figure out what the original commit was that introduced this code, and then
seeing your own name on that commit?

Yeah, this piece of code of has a small error that basically invalidated the
whole purpose of the changes made. Happily, the error was to let through
everything, which is still *ok*, but the desired result was the filter
certain things.

The error was: I created a function to remove "pos" fields from "sounds"
data when adding that data to a list. The first part of the block was about
filtering out "sounds" data that didn't match the word or the forms of the
word we were being processed, and the *second* part of the block was about
filtering for wrong "pos" data and then removing the "pos" sections.. But I
used the pos-removing function in the first part of the block, which meant
there were no pos-sections to compare in the second. Because there was
not pos-data, all the sounds were let through.