Extract the rest tables in the "fr-conj-*" templates
There are two templates in the "fr-conj" template, one of
them("fr-conj/tableau-temps-simple") expands to HTML table and the
other("fr-conj/Tableau-composé") expands to wikitext table.

There is still one problem, some form texts are missing white space
between the words. For example: "tu as lancé" is expanded to "tuas
lancé". But this is not fixable in extractor code.