Spanish edition: Change pydantic classes
Fields with singular names should only contain singular
items and not lists; in this case, I've opted to change
the type (list -> Optional[str]) instead of changing the
names ("ipa" -> "ipas").

We never use "tag" as a field, although it's not incorrect;
in this case, I changed the name to "tags" and left it as a
list of strings. The strings should be space separated
tags, this field should not be used for arbitrary ad-hoc 'tags'!