Spanish edition: pronunciation test changes
Changed excpected output to match current output.

One change to a test: one test (no name because these are in a list)
had an "audio2" parameter without having any sort of other "___2" besides it,
so the outputted data was malformed. This was because
this was copypasted from an article where there was ___2 data,
but that data was removed to make the test material smaller.

That is, `{{pron-graf|leng=en|pron=británico|audio2=En-uk-direction.ogg|`
resulted in:
```
{"tags": ["británico"]},
{"tags": ["británico"],
 "audio": ...},
...
```

because a "fono2" was removed from the original material. The original
data had a "fono" and "fono2" field, which filled up the above
dicts more sensibly.