Map out data types for word data
Currently these are TypedDicts with `total=False`, but
it is possible in Python 3.9 to have TypedDicts with
mixed required and optional fields by using inheritance;
one parent TypedDict has the required fields (`total=True`),
the other doesn't, and the child TypedDict class has
mixed requirements. This can be done later.