Remove parse_linkage_template
This code has been unreachable for a couple of years,
with an XXX comment about removing it; it's pretty hard to
figure out if the changes Tatu was meaning to make actually
were made, but seeing as how things have been running smoothly
even when parse_linkage_template has been unreachable all
this time, I opted to just remove the unused code.

I've also commented out some data used by parse_linkage_template
in linkage_template_mappings (a list of lists that should have
been a dict?), and moved the linkage template names into a new
template_linkages set to be used in the one place that still
had a reference to it.