Es-edition: Remove kludges, use substring "whitelist"
Really, all the previous kludges in the previous commit are
unnecessary if I'd just realized we only want to keep
etymology template data for *etymology templates*, not
just any template. So a simple substring check should suffice.
It is unlikely that there are any conflicting templates
with "etim" in their name, but possible.