Add clean_node_fn parameter to clean_node()
For convenience, to allow easier access to the functionality
of clean_node without having to fork it, I've added an optional
parameter to clean_node that receives a function to handle and
process WikiNodes. Previously, there was a default function
defined inside clean_nodes that was passed on to Wtp.node_to_html(),
but clean_node() is currently the most complete "cleaning function"
we have (should probably be forked or moved (to a new package??)),
so might as well.

Shouldn't break anything.