Inline images should print alt= text
If an image is inline (it doesn't have any of a set
of arguments that, like 'right' or 'thumb' etc.), then
its `alt=` text should be printed when returned from
clean_value.

To make it clearer that this is an `alt` text, and to
make post-processing easier, we add `[Alt: ` + `]` around
the text.