Change to Fr inflection stuff
Because test_invalid_ipa in test_fr_inflection failed
due to changes to how the URL was handled, I've changed
the way no-ipa is detected by removing a space.

I also believe the function should return an empty
list of strings instead of an empty string; didn't matter
here, because the only test is whether len(ret) == 0, but
I believe it's incorrect based on the type annotations?

The backslashes before and after "Prononciation ?" seem
strange?

Someone needs to check this.