Changes to classify_desc heuristics
"He" is definitely English in some cases. Added
a bit more heuristics and a new `potentially_english_
words` used in the blacklist but later used to check
that if all rejected words are in it and the
number of rejected words is less than 50%, then
accept as `english`.

Added `co`, `coi` and `collocation` as example templates
because they look like example templates and quack
like example templates.