Use example template args to determine example
Previously, we would expand everything with clean_node
and use heuristics to check whether it was an example
or not.

However, seeing as how we have access to template
arguments, we can bypass some of these heuristics
early by checking if the text we're checking has
one template, from the example templates set,
with arguments that are the same as what is part
of the clean_node output. In this case, we can just
assume this is an example.