Attempt at fixing issues with level 3 sections
Chinese sections often have up to three level 3
sections one after another; a Glyph Origin,
Etymology and Pronunciation. Later sections
override the previous (or rather, the previous)
is ignored; in the previous commit, we combined
Etymology sections with an immediately preceding
(without subsections, important) Level 3 section
(Glyph Origin), which worked fine in simple examples,
but when you have Pronunciation sections (taking
the slot of an Etymology on Level 3) there was
still no etymology data.

This commit is trying to put the Pronunciation section
on level 4 so that it's the child of the Etymology
section, but it didn't work out. This commit is
mainly for history so that I can wipe the previous
changes to fix_subtitle hierarchy()