Introduce a new level between Etymology and POS
This is needed for Chinese entries with structures like:

===Glyph Origin===
===Etymology===
===Pronunciation 1===
====POS Section====
===Pronunciation 2===
====POS Section====

Glyph Origin and Etymology are merged in fix_subtitle_hierarchy
(previous commit), and Pronunciation sections are moved to
Level 4, while all other sections under that are move one
step further, also; Levels 2 and 3 (Language and Etymology)
are left alone, and happily they're also the only truly
"meaningful" sections; The code doesn't really check for
`NodeKind.LEVEL[3-6]` anywhere.

The inbetween stage for Pronunciation only involves acting
as a bridge between push_etym() and push_pos(). 99% of the time
(or maybe a slightly lower percentage, there are a lot of
Chinese entries) it doesn't really do anything.