Fixing?? issues with switch over to parsing level 4's
Chinese word entries sometimes have several level 3
sections after each other like this:

```
===Glyph Origin===
===Etymology===
===Pronunciation 1===
===Pronunciation 2===
```

where these should all be part of the same thing.

Previous changes attempted to fix this by inserting a
virtual Level 4 into the mix, by changing Pronunciation
sections to level 4 so they would be hierarchically
under Etymology sections (and Etymology sections
are combined under one).

The last commit didn't work perfectly, but I think I got it
right this time... At least it looks like it, and
tests pass.

It took me ages. I just couldn't get it right.

Anyhow, currently it works like this:

There's POS data and etym data as previously, but they
are bridged by level_four data. Usually the level_four
data is empty, except when a pronunciation section is found,
in which case we flip a flag to true that means that
when we parse stuff and add the data to something, that
data is added to the level_four data instead of `etym`.

`etym` is not etymological data, it's data for an
Etymology *section*, so it's everything under an
Etymology title, unless there's a Pronunciation title,
in which case `etym` should contain mostly etymology
and general data (like categories plucked from the
etymology section). The name is a bit confusing, but
changing to "level_three" is annoyingly abstract too.