Remove unused code
We used to have a "conjugation" field, but not anymore.

The piece of code related to "topics" has not been
functional for a long time, if ever; the intent seems
to have been to copy topic tags between senses, but here
we process whole words. Check the json output of en.wiktionary,
and instances of "inaccurate" (which the removed code would
have introduced) are rare and at a quick glance, none of
the instances were generated by this code.

The code and logic is so old that it's best to just remove it.
If we want to introduce this functionality (again? if it
was ever there), it's trivial to recreate.