[en] extract forms data and literal meaning from "zh-forms" template
The code is modified from the zh edition code. The new function adds
two tags("Simplified Chinese", "Traditional Chinese") and a new
"literal_meaning" JSON field. The literal meaning data is commonly
provided in Chengyu pages.

Data are added to `base_data` because "zh-forms" is above POS sections
and should be included to all the following POS sections data.

Since the code are mostly the same as the zh edition code, tests are
not added.

Example pages:
- https://en.wiktionary.org/wiki/未雨綢繆  (Chengyu, has `lit` parameter)
- https://en.wiktionary.org/wiki/僥倖  (has two POS sections)

GitHub issue #676