Add `info_templates` field, handle `Template:+obj`
`info_templates` is similar to etym_templates and
head_templates, but can also have an `extra_data`
field with more specific data relating to the
template in question.

For `+obj`, a vague template that generates text
like `[+accusative]` often meaning 'used with
accusative objects' or `[+dative = means this things]`
meaning "if this is with dative it means `this
thing`", but sometimes not, this extra data consists
of a `meaning` field from the `means=` parameter,
and `tags` and `words` for parsing the output that
isn't the `means=` field.